requirements.txt
\
torch
torchvision
transformers
datasets
scikit-learn
Pillow
accelerate 